Feat/timeline updates

+++ src/repositories/postsRepository.js
@@ -1,4 +1,4 @@
-import db from "../database/database.js";
+import db from '../database/database.js';
 
 const postInsertion = ({ url, content, userId, metadataId }) => {
   return db.query(
@@ -8,20 +8,18 @@ const postInsertion = ({ url, content, userId, metadataId }) => {
 };
 
 const hashtagInsertion = (hashtag) => {
-  return db.query("INSERT INTO hashtags (name) VALUES ($1) RETURNING id;", [
-    hashtag,
-  ]);
+  return db.query('INSERT INTO hashtags (name) VALUES ($1) RETURNING id;', [hashtag]);
 };
 
 const selectHashtag = (hashtag) => {
-  return db.query("SELECT * FROM hashtags WHERE name = $1", [hashtag]);
+  return db.query('SELECT * FROM hashtags WHERE name = $1', [hashtag]);
 };
 
 const hashtagsPostsInsertion = ({ postId, hashtagId }) => {
-  return db.query(
-    `INSERT INTO "hashtagsPosts" ("postId", "hashtagId") VALUES ($1, $2);`,
-    [postId, hashtagId]
-  );
+  return db.query(`INSERT INTO "hashtagsPosts" ("postId", "hashtagId") VALUES ($1, $2);`, [
+    postId,
+    hashtagId,
+  ]);
 };
 
 async function insertLinkMetadata({ image, title, description }) {
@@ -45,7 +43,7 @@ async function insertPostVisits({ postId }) {
   );
 }
 
-async function getRecentPosts({ limit }) {
+async function getPostsWithLimitAndOffset({ userId, limit }) {
   return db.query(
     `SELECT "userWhoRepost", "nameUserWhoRepost", "id", "url", "content" , "user", "metadata", "usersWhoLiked", "visitCount", "hashtagsList", "createdAt" FROM(SELECT
       "p"."createdAt" AS "createdAt",
@@ -83,91 +81,8 @@ async function getRecentPosts({ limit }) {
         LEFT JOIN visits "v" ON "v"."postId" = "p"."id"
       LEFT JOIN reposts "r" ON "r"."postId" = "p"."id"
       LEFT JOIN users "u2" ON "r"."userId" = "u2"."id"
-      UNION ALL
-      SELECT
-      "r"."createdAt" AS "createdAt",
-      "r"."userId" AS "userWhoRepost",
-      "u2".name AS "nameUserWhoRepost",
-        "p"."id" AS "id",
-        "p"."url",
-        "p"."content",
-        json_build_object('name', "u"."name", 'email', "u"."email", 'profilePic', "u"."profilePic", 'id', "u"."id") AS "user",
-        json_build_object('image', "m"."image", 'title', "m"."title", 'description', "m"."description") AS "metadata",
-        ARRAY(
-          SELECT
-              json_build_object('name', "l_u"."name", 'email', "l_u"."email")
-          FROM
-              posts "l_p"
-              LEFT JOIN likes "l_l" ON l_l."postId" = "l_p"."id"
-              JOIN users "l_u" ON "l_u"."id" = "l_l"."userId"
-          WHERE "l_l"."postId" = "p"."id"
-          ORDER BY "l_l"."createdAt" DESC
-        ) AS "usersWhoLiked",
-        COALESCE ("v"."count", 0) AS "visitCount",
-        ARRAY(
-          SELECT
-            json_build_object('id',"h_h"."id", 'name', "h_h"."name")
-          FROM
-            "posts" "h_p"
-            JOIN "hashtagsPosts" "h_hp" ON "h_p"."id" = "h_hp"."postId"
-            JOIN "hashtags" "h_h" ON "h_hp"."hashtagId" = "h_h"."id"
-          WHERE "h_p"."id" = "p"."id"
-        ) AS "hashtagsList"
-      FROM
-        posts "p"
-        JOIN users "u" ON "p"."userId" = "u"."id"
-        JOIN metadata "m" ON "p"."metadataId" = "m"."id"
-        LEFT JOIN visits "v" ON "v"."postId" = "p"."id"
-      RIGHT JOIN reposts "r" ON "r"."postId" = "p"."id"
-      LEFT JOIN users "u2" ON "r"."userId" = "u2"."id")
-    AS results
-      ORDER BY "createdAt" DESC
-      LIMIT $1;
-	`,
-    [limit]
-  );
-}
-
-async function getRecentPostsTimeline({ userId, limit }) {
-  return db.query(
-    `SELECT "userWhoRepost", "nameUserWhoRepost", "id", "url", "content" , "user", "metadata", "usersWhoLiked", "visitCount", "hashtagsList", "createdAt" FROM(SELECT
-      "p"."createdAt" AS "createdAt",
-      null AS "userWhoRepost",
-      null AS "nameUserWhoRepost",																																	 
-        "p"."id" AS "id",
-        "p"."url",
-        "p"."content",
-        json_build_object('name', "u"."name", 'email', "u"."email", 'profilePic', "u"."profilePic", 'id', "u"."id") AS "user",
-        json_build_object('image', "m"."image", 'title', "m"."title", 'description', "m"."description") AS "metadata",
-        ARRAY(
-          SELECT
-              json_build_object('name', "l_u"."name", 'email', "l_u"."email")
-          FROM
-              posts "l_p"
-              LEFT JOIN likes "l_l" ON l_l."postId" = "l_p"."id"
-              JOIN users "l_u" ON "l_u"."id" = "l_l"."userId"
-          WHERE "l_l"."postId" = "p"."id"
-          ORDER BY "l_l"."createdAt" DESC
-        ) AS "usersWhoLiked",
-        COALESCE ("v"."count", 0) AS "visitCount",
-        ARRAY(
-          SELECT
-            json_build_object('id',"h_h"."id", 'name', "h_h"."name")
-          FROM
-            "posts" "h_p"
-            JOIN "hashtagsPosts" "h_hp" ON "h_p"."id" = "h_hp"."postId"
-            JOIN "hashtags" "h_h" ON "h_hp"."hashtagId" = "h_h"."id"
-          WHERE "h_p"."id" = "p"."id"
-        ) AS "hashtagsList"
-      FROM
-        posts "p"
-        JOIN users "u" ON "p"."userId" = "u"."id"
-        JOIN metadata "m" ON "p"."metadataId" = "m"."id"
-        LEFT JOIN visits "v" ON "v"."postId" = "p"."id"
-      LEFT JOIN reposts "r" ON "r"."postId" = "p"."id"
-      LEFT JOIN users "u2" ON "r"."userId" = "u2"."id"
-	    LEFT JOIN followers f ON f."followerId" = u.id
-	    WHERE f."userId" = $1
+	    LEFT JOIN followers f ON f."userId" = u.id
+	    WHERE f."followerId" = $1 OR p."userId" = $1
       UNION ALL
       SELECT
       "r"."createdAt" AS "createdAt",
@@ -208,8 +123,9 @@ async function getRecentPostsTimeline({ userId, limit }) {
     AS results
       ORDER BY "createdAt" DESC
       LIMIT $2;
+      OFFSET $3;
 	`,
-    [userId, limit]
+    [userId, limit, offset]
   );
 }
 
@@ -218,38 +134,23 @@ async function getPostById(postId) {
 }
 
 async function updateContentPost(postId, content) {
-  return db.query(`UPDATE posts SET content = $1 WHERE posts.id = $2;`, [
-    content,
-    postId,
-  ]);
+  return db.query(`UPDATE posts SET content = $1 WHERE posts.id = $2;`, [content, postId]);
 }
 
 async function getUserLikeOnPostById({ postId, userId }) {
-  return db.query(
-    `SELECT * FROM likes WHERE "postId" = $1 AND "userId" = $2;`,
-    [postId, userId]
-  );
+  return db.query(`SELECT * FROM likes WHERE "postId" = $1 AND "userId" = $2;`, [postId, userId]);
 }
 
 async function likePostById({ postId, userId }) {
-  return db.query(`INSERT INTO likes ("userId", "postId") VALUES ($1, $2);`, [
-    userId,
-    postId,
-  ]);
+  return db.query(`INSERT INTO likes ("userId", "postId") VALUES ($1, $2);`, [userId, postId]);
 }
 
 async function dislikePostById({ postId, userId }) {
-  return db.query(`DELETE FROM likes WHERE "postId" = $1 AND "userId" = $2;`, [
-    postId,
-    userId,
-  ]);
+  return db.query(`DELETE FROM likes WHERE "postId" = $1 AND "userId" = $2;`, [postId, userId]);
 }
 
 const deletePostById = ({ postId, userId }) => {
-  return db.query(`DELETE FROM posts WHERE "id" = $1 AND "userId" = $2;`, [
-    postId,
-    userId,
-  ]);
+  return db.query(`DELETE FROM posts WHERE "id" = $1 AND "userId" = $2;`, [postId, userId]);
 };
 
 async function getRepostByPostId(postId) {
@@ -261,24 +162,19 @@ async function getRepostByUserId(userId) {
 }
 
 async function getRepostByUserIdandPostId(userId, postId) {
-  return db.query(
-    `SELECT * FROM reposts WHERE "userId" = $1 AND "postId" = $2;`,
-    [userId, postId]
-  );
+  return db.query(`SELECT * FROM reposts WHERE "userId" = $1 AND "postId" = $2;`, [userId, postId]);
 }
 
 async function insertRepost({ userId, postId }) {
-  return db.query(`INSERT INTO reposts ("userId", "postId") VALUES ($1,$2);`, [
-    userId,
-    postId,
-  ]);
+  return db.query(`INSERT INTO reposts ("userId", "postId") VALUES ($1,$2);`, [userId, postId]);
 }
 
 const insertCommentOnPost = ({ postId, userId, content }) => {
-  return db.query(
-    `INSERT INTO comments ("postId", "userId", content) VALUES ($1, $2, $3);`,
-    [postId, userId, content]
-  );
+  return db.query(`INSERT INTO comments ("postId", "userId", content) VALUES ($1, $2, $3);`, [
+    postId,
+    userId,
+    content,
+  ]);
 };
 
 const getCommentsById = ({ postId }) => {
@@ -306,7 +202,7 @@ export {
   postInsertion,
   insertLinkMetadata,
   insertPostVisits,
-  getRecentPosts,
+  getPostsWithLimitAndOffset,
   hashtagInsertion,
   hashtagsPostsInsertion,
   selectHashtag,
@@ -322,5 +218,4 @@ export {
   getRepostByUserIdandPostId,
   insertCommentOnPost,
   getCommentsById,
-  getRecentPostsTimeline
 };
